Update to 1.21.4
Didn't test a single thing so hopefully it works. It builds, there was a
lot of API changes in the various plugins it brings in so hopefully I'm
doing everything right. The async calls in the Coins plugin is my most
questionable area but oh well it should work fine?